chore: initially tease transformer-factory out from transformGraphQLSchema
