+++ packages/amplify-provider-awscloudformation/src/graphql-transformer/index.ts
@@ -1,1 +1,2 @@
-export * from './user-defined-slots';
+export { SLOT_NAMES, parseUserDefinedSlots } from './user-defined-slots';
+export { transformGraphQLSchema } from './transform-graphql-schema';
